raspi-image: Use losetup for handling loop devices
Use losetup instead of partx to attach and detach image files to loop
devices. It seems better suitable and detaches images more reliably
than partx. While at it, use globals for the mountpoint and loop device
name so that we don't have to carry them around and can handle the
cleanup on error conditions.

Signed-off-by: Juerg Haefliger <juergh@gmail.com>